# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Mini-Orca API
  description: |
    Mini-Orca is a local agents orchestrator for mini tasks. It uses multiple specialized LLM agents
    (Planner, Coder, Tester, Reviewer) to plan, code, test, and review project tasks autonomously
    through a phase-based workflow.

    ## Versioning Strategy

    Mini-Orca follows [Semantic Versioning](https://semver.org/) (SemVer) for its API:

    - **MAJOR version (v2)**: Incremented for incompatible API changes.
      - v2.0.0 introduced the `/api/render/*` HTMX endpoints, project management APIs, and updated session lifecycle.
      - v1.x used a different session model and lacked project management.
    - **MINOR version (x.1)**: Incremented for backward-compatible additions (new endpoints, fields).
    - **PATCH version (x.y.1)**: Incremented for backward-compatible bug fixes.

    ## Version Detection

    Access the current API version via:
    - `GET /health` → `{"status":"ok","version":"2.0.0"}`
    - `GET /status` → `{"status":"running","version":"2.0.0",...}`

    ## Authentication

    Version 2.0 does not require authentication. All endpoints are publicly accessible.

    ## Rate Limiting

    No rate limits are currently configured.

    ## Error Handling

    All errors follow a consistent JSON format:

    ```json
    {
      "type": "bad_request",
      "message": "Technical error description for developers",
      "user_message": "User-friendly error message",
      "code": 400
    }
...

...
servers:
...
paths:
...
  /api/sessions:
    post:
      tags:
        - Sessions
      summary: Create session
      description: Creates a new agent session for autonomous task execution.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreateRequest"
            example:
              goal: "Implement user authentication module"
              project_path: "/path/to/project"
              project_type: "go"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
              example:
                id: "session-1719000000000000000"
                goal: "Implement user authentication module"
                project_path: "/path/to/project"
                project_type: "go"
                current_phase: "planning"
                status: "pending"
                created_at: "2024-07-01T12:00:00Z"
                updated_at: "2024-07-01T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags:
        - Sessions
      summary: List sessions
      description: Returns a list of all sessions with their summaries.
      operationId: listSessions
      responses:
        "200":
          description: Session list retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
              example:
                sessions:
                  - id: "session-1719000000000000000"
                    goal: "Implement user authentication module"
                    current_phase: "coding"
                    status: "running"
                    created_at: "2024-07-01T12:00:00Z"
                total: 1
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get session status
      description: Returns the current status of a specific session.
      operationId: getSessionStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        "200":
          description: Session details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sessions/{id}/start:
    post:
      tags:
        - Sessions
      summary: Start session
      description: Transitions a session from pending to running and begins the planning phase.
      operationId: startSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionActionResponse"
              example:
                status: started
                session_id: "session-1719000000000000000"
                phase: "planning"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sessions/{id}/pause:
    post:
      tags:
        - Sessions
      summary: Pause session
      description: Pauses a running session.
      operationId: pauseSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        "200":
          description: Session paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionActionResponse"
              example:
                status: paused
                session_id: "session-1719000000000000000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sessions/{id}/resume:
    post:
      tags:
        - Sessions
      summary: Resume session
      description: Resumes a paused session.
      operationId: resumeSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        "200":
          description: Session resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionActionResponse"
              example:
                status: resumed
                session_id: "session-1719000000000000000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sessions/{id}/stop:
    post:
      tags:
        - Sessions
      summary: Stop session
      description: Stops a running session, transitioning it to completed state.
      operationId: stopSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        "200":
          description: Session stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionActionResponse"
              example:
                status: stopped
                session_id: "session-1719000000000000000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    SessionCreateRequest:
      type: object
      required:
        - goal
        - project_path
      properties:
        goal:
          type: string
          description: High-level goal of the session
          example: "Implement user authentication module"
        project_path:
          type: string
          description: Filesystem path to the project directory
          example: "/path/to/project"
        project_type:
          type: string
          description: Project type override (auto-detected if omitted)
          enum:
            - go
            - python
            - typescript
            - rust
            - java
            - kotlin
            - unknown
          example: "go"
    SessionResponse:
      type: object
      required:
        - id
        - goal
        - project_path
        - project_type
        - current_phase
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique session identifier
          example: "session-1719000000000000000"
        goal:
          type: string
          description: The session goal
          example: "Implement user authentication module"
        project_path:
          type: string
          description: Path to the project
          example: "/path/to/project"
        project_type:
          type: string
          description: Detected or specified project type
          example: "go"
        current_phase:
          type: string
          description: Current phase of the session
          enum:
            - planning
            - coding
            - testing
            - review
          example: "planning"
        status:
          type: string
          description: Current session status
          enum:
            - pending
            - running
            - paused
            - completed
            - cancelled
          example: "pending"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-07-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2024-07-01T12:00:00Z"
        error:
          type: string
          description: Optional error message if session failed
          nullable: true
          example: ""
    SessionListResponse:
      type: object
      required:
        - sessions
        - total
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
          description: List of session summaries
        total:
          type: integer
          description: Total number of sessions
          example: 1
    SessionSummary:
      type: object
      required:
        - id
        - goal
        - current_phase
        - status
        - created_at
      properties:
        id:
          type: string
          description: Unique session identifier
          example: "session-1719000000000000000"
        goal:
          type: string
          description: The session goal
          example: "Implement user authentication module"
        current_phase:
          type: string
          description: Current phase
          enum:
            - planning
            - coding
            - testing
            - review
          example: "coding"
        status:
          type: string
          description: Current status
          enum:
            - pending
            - running
            - paused
            - completed
            - cancelled
          example: "running"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-07-01T12:00:00Z"
    SessionActionResponse:
      type: object
      required:
        - status
        - session_id
      properties:
        status:
          type: string
          description: Action result
          example: started
        session_id:
          type: string
          description: Session identifier
          example: "session-1719000000000000000"
        phase:
          type: string
          description: Phase name (only on start)
          enum:
            - planning
            - coding
            - testing
            - review
          nullable: true
          example: "planning"

    # ─── Gates ─────────────────────────────────────────────────────────
...
    Error:
      type: object
      required:
        - type
        - message
        - user_message
        - code
      properties:
        type:
          type: string
          description: Error category
          enum:
            - internal
            - not_found
            - bad_request
            - unauthorized
            - forbidden
            - conflict
          example: bad_request
        message:
          type: string
          description: Technical error description for developers
          example: "session: goal is required"
        user_message:
          type: string
          description: User-friendly error message
          example: "A session goal must be provided."
        code:
          type: integer
          description: HTTP status code
          example: 400
  responses:
    BadRequest:
      description: Bad request - invalid parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: bad_request
            message: "session: goal is required"
            user_message: "A session goal must be provided."
            code: 400
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: not_found
            message: "session: \"abc\" not found"
            user_message: "The requested resource was not found."
            code: 404
    Conflict:
      description: Conflict - invalid state transition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: conflict
            message: "The session must be in pending state to start."
            user_message: "A conflict occurred with the current state of the resource."
            code: 409
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: internal
            message: "Failed to update session status."
            user_message: "An internal server error occurred. Please try again later."
            code: 500